AWS-Deployment-Automation/WordpressEcsConstruct/requirements.txt
-e .
aws-cdk.aws-rds
boto3